Added phpstan

+++ src/Strategies/CartPercentDiscountStrategy.php
@@ -20,6 +20,7 @@ class CartPercentDiscountStrategy extends DiscountStrategy implements DiscountSt
         if (app(CouponServiceContract::class)->cartItemIsExcludedFromCoupon($this->coupon, $cartItem)) {
             return 0;
         }
-        return round($this->coupon->amount * $cartItem->buyable->getBuyablePrice() / 100, 0);
+        // @phpstan-ignore-next-line
+        return (int) round($this->coupon->amount * $cartItem->buyable->getBuyablePrice() / 100, 0);
     }
 }
